Example input census data for package use

This is example forest census data to be analyzed with this package.

census_1_ex

Format

A tibble

ID

Tree identification number. This identifies an individual tree and can be used to connect trees between the two censuses.

sp

Species of the individual

gx

x-coordinate meters from reference point

gy

y-coordinate meters from reference point

date

Date the stem was measured

codes

Code for additional information on the stem: M means the main stem of the individual tree and R means the stem was lost, but the tag was moved to another stem greater than DBH cutoff, this stands for resprout.

dbh

Diameter at breast height of the tree in cm

See also

Examples

library(dplyr)
library(stringr)

growth_ex <-
  compute_growth(
    census_1 = census_1_ex %>%
      mutate(sp = to_any_case(sp) %>% factor()),
    census_2 = census_2_ex %>%
      filter(!str_detect(codes, "R")) %>%
      mutate(sp = to_any_case(sp) %>% factor()),
    id = "ID"
  )