Example focal versus comp data frame

This is an example focal versus comp data frame. The rows are focal trees which are repeated for all competitor trees within a specified distance from them. In this case that distance is 1. This is the focal versus comp for growth_spatial_ex.

focal_vs_comp_ex

Format

A tibble::tbl_df:

focal_ID

Tree identification number for the focal tree

focal_sp

Species of the focal tree

dbh

Diameter at breast height of the focal tree at the first census

foldID

The CV-fold that the focal tree is in

geometry

The point location of the focal tree

growth

The average annual growth of the focal tree between censuses

comp

A list-column: characteristics of the relevant competitor trees

The comp list-column contains tibble::tbl_dfs with columns:

comp_ID

Tree identification number for the competitor tree

dist

The distance between the focal and comp tree, this will be less than the max distance specified.

comp_sp

Species of the comp tree

comp_x_var

Numerical variable associated with comp tree

See also

Examples

comp_bayes_lm_ex <- focal_vs_comp_ex %>%
  comp_bayes_lm(prior_param = NULL, run_shuffle = FALSE)