Phylogenic groupings and trait based clustering of various tree species

A date frame mapping the species codes to their common names, scientific names, and families. This also includes a trait-based clustering of the species.

species_bw

Format

A data frame with 46 rows and 6 variables:

sp

The code for the species. Link to census_2008_bw and census_2014_bw with sp variable.

genus

Genus

species

Species epithet

latin

Scientific name

family

Family

trait_group

Clustering of species based on three traits rather than their evolutionary relationships. The traits are specific leaf area, maximum height, and wood density

Source

For more information on trait clustering see Allen and Kim 2020 "A permutation test and spatial cross-validation approach to assess models of interspecific competition between trees." doi: 10.1371/journal.pone.0229930 PLOS One 15: e0229930.

See also

Examples

library(dplyr)

# Original 2008 census data
census_2008_bw

#> # A tibble: 27,193 x 8
#>    treeID stemID sp              gx    gy   dbh date       codes
#>     <dbl>  <dbl> <chr>        <dbl> <dbl> <dbl> <date>     <chr>
#>  1      1      1 white_oak      8.7  108.  41.2 2008-07-01 M    
#>  2      2      1 serviceberry  10.8  109.   5.9 2008-07-01 M    
#>  3      3      1 black_cherry  11.6  106.  45.2 2008-07-01 M    
#>  4  14951      1 black_cherry  12    107    3.8 2008-07-01 M    
#>  5      4      1 red_maple     12.5  108.   4.1 2008-07-01 M    
#>  6      5      1 red_maple     12.5  106.   4.6 2008-07-01 M    
#>  7      6      1 red_maple     13.3  107.   8.3 2008-07-01 M    
#>  8      7      1 red_maple     14.9  106.   6.1 2008-07-01 M    
#>  9      8      1 red_maple     15.6  109    7.7 2008-07-01 M    
#> 10      9      1 red_maple     16.5  108.   4   2008-07-01 M    
#> # … with 27,183 more rows

# 2008 census data with additional species information
census_2008_bw %>%
  left_join(species_bw, by = "sp")

#> # A tibble: 27,193 x 13
#>    treeID stemID sp        gx    gy   dbh date       codes genus  species latin 
#>     <dbl>  <dbl> <chr>  <dbl> <dbl> <dbl> <date>     <chr> <chr>  <chr>   <chr> 
#>  1      1      1 white…   8.7  108.  41.2 2008-07-01 M     querc… alba    querc…
#>  2      2      1 servi…  10.8  109.   5.9 2008-07-01 M     amela… arborea amela…
#>  3      3      1 black…  11.6  106.  45.2 2008-07-01 M     prunus seroti… prunu…
#>  4  14951      1 black…  12    107    3.8 2008-07-01 M     prunus seroti… prunu…
#>  5      4      1 red_m…  12.5  108.   4.1 2008-07-01 M     acer   rubrum  acer_…
#>  6      5      1 red_m…  12.5  106.   4.6 2008-07-01 M     acer   rubrum  acer_…
#>  7      6      1 red_m…  13.3  107.   8.3 2008-07-01 M     acer   rubrum  acer_…
#>  8      7      1 red_m…  14.9  106.   6.1 2008-07-01 M     acer   rubrum  acer_…
#>  9      8      1 red_m…  15.6  109    7.7 2008-07-01 M     acer   rubrum  acer_…
#> 10      9      1 red_m…  16.5  108.   4   2008-07-01 M     acer   rubrum  acer_…
#> # … with 27,183 more rows, and 2 more variables: family <chr>,
#> #   trait_group <chr>